requirements.txt
alpaca-py>=0.20.0
pandas>=2.0.0
python-dotenv>=1.0.0
plotly>=5.0.0
ccxt>=4.0.0
